+++ packages/cloud_firestore/lib/src/query.dart
@@ -50,7 +50,9 @@ class Query {
 
   /// Notifies of query results at this location
   // TODO(jackson): Reduce code duplication with [DocumentReference]
-  Stream<QuerySnapshot> snapshots() {
+  Stream<QuerySnapshot> snapshots(
+      {MetadataChanges metadataChanges = MetadataChanges.exclude}) {
+    assert(metadataChanges != null);
     Future<int> _handle;
     // It's fine to let the StreamController be garbage collected once all the
     // subscribers have cancelled; this analyzer warning is safe to ignore.
@@ -64,6 +66,7 @@ class Query {
             'path': _path,
             'isCollectionGroup': _isCollectionGroup,
             'parameters': _parameters,
+            'metadataChanges': _getMetadataChangesString(metadataChanges),
           },
         ).then<int>((dynamic result) => result);
         _handle.then((int handle) {